ucd update

+++ lib/tasks/import/ucd/ucd.rake
@@ -183,6 +183,7 @@ namespace :tw do
 #end
 
         invalid_relationship_remove
+        invalid_relationship_remove
 
         print "\n\n !! Success. End time: #{Time.now} \n\n"
 
@@ -760,6 +761,7 @@ namespace :tw do
               c.genus = TaxonName.find(origgen) unless origgen.nil?
               c.subgenus = TaxonName.find(origsubgen) unless origsubgen.nil?
               c.species = taxon
+              c.verbatim_name = c.get_full_name
               c.save
               if c.id.nil?
                 c1 = Combination.match_exists?(c.get_full_name, genus: c.genus.try(:id), subgenus: c.subgenus.try(:id), species: c.species.try(:id))
@@ -842,6 +844,7 @@ namespace :tw do
               c.subgenus = TaxonName.find(origsubgen) unless origsubgen.nil?
               c.species = TaxonName.find(origspecies) unless origspecies.nil?
               c.subspecies = taxon
+              c.verbatim_name = c.get_full_name
               c.save
               if c.id.nil?
                 c1 = Combination.match_exists?(c.get_full_name, genus: c.genus.try(:id), subgenus: c.subgenus.try(:id), species: c.species.try(:id), subspecies: c.subspecies.try(:id))
@@ -2093,18 +2096,22 @@ namespace :tw do
 #=begin
         j = 0
         print "\nHandling Invalid relationships: synonyms of synonyms\n"
-        tr = TaxonNameRelationship.where(project_id: $project_id).with_type_base('TaxonNameRelationship::Iczn::Invalidating')
+        tr = TaxonNameRelationship.where(project_id: $project_id).with_type_array(TAXON_NAME_RELATIONSHIP_NAMES_SYNONYM)
         tr.each do |t|
           j += 1
+
           print "\r#{j}    Fixes applied: #{fixed}   "
           s = t.subject_taxon_name
           o = t.object_taxon_name
           sval = s.valid_taxon_name
-          #next unless s.name == 'mala' || s.name == 'Narayana' || s.name == 'subaenea' || s.name == 'subaeneus'
+#          next unless s.name == 'hispanicus' || s.name == 'hispanica'
+#          byebug
+
           if o.rank_string =~ /Family/
             if o.id != sval.id && o.cached_primary_homonym_alternative_spelling == sval.cached_primary_homonym_alternative_spelling
               t.object_taxon_name = sval
               t.save
+              s.save
               fixed += 1
             end
             if s.cached_primary_homonym_alternative_spelling != o.cached_primary_homonym_alternative_spelling && s.origin_citation.nil?
@@ -2112,15 +2119,18 @@ namespace :tw do
                 if !p.origin_citation.nil?
                   t.object_taxon_name = p
                   t.save
+                  s.save
                   fixed += 1
                 end
               end
 
             end
           else
+            #byebug if j == 7216
             if o.id != sval.id && o.cached_secondary_homonym_alternative_spelling == sval.cached_secondary_homonym_alternative_spelling
               t.object_taxon_name = sval
               t.save
+              s.save
               fixed += 1
             end
             if s.cached_secondary_homonym_alternative_spelling != o.cached_secondary_homonym_alternative_spelling && s.origin_citation.nil?
@@ -2128,6 +2138,7 @@ namespace :tw do
                if !p.origin_citation.nil?
                   t.object_taxon_name = p
                   t.save
+                  s.save
                   fixed += 1
                 end
               end
@@ -2137,36 +2148,36 @@ namespace :tw do
         end
 #end
 
-
-
       print "\nHandling Invalid relationships: synonyms to combinations\n"
       tr = TaxonNameRelationship.where(project_id: $project_id).with_type_string('TaxonNameRelationship::Iczn::Invalidating')
         tr.each do |t|
           i += 1
           print "\r#{i}    Fixes applied: #{fixed}    Combinations created: #{combinations}"
-
           if t.citations.empty?
             s = t.subject_taxon_name
             svalid = s.cached_valid_taxon_name_id
             o = t.object_taxon_name
             shas = s.cached_secondary_homonym_alternative_spelling
+            r = TaxonNameRelationship.where(project_id: $project_id, object_taxon_name_id: s.id).with_type_base('TaxonNameRelationship::Iczn::Invalidating')
+            r2 = TaxonNameRelationship.where(project_id: $project_id, subject_taxon_name_id: s.id).with_type_base('TaxonNameRelationship::Iczn::Invalidating').count
 
-            #next unless s.name == 'mala' || s.name == 'Narayana' || s.name == 'subaenea' || s.name == 'subaeneus'
-            #byebug
-            if s.taxon_name_classifications.empty?
+#            next unless s.name == 'hispanicus' || s.name == 'hispanica'
+#            byebug
+            if s.taxon_name_classifications.empty? && r.empty?
               t.destroy
               s.save
-                if o.rank_string =~ /Family/ && s.cached_primary_homonym_alternative_spelling == o.cached_primary_homonym_alternative_spelling
+                if o.rank_string =~ /Family/ && s.cached_primary_homonym_alternative_spelling == o.cached_primary_homonym_alternative_spelling && r2 == 1
                     fixed += 1
                     TaxonNameRelationship.create!(subject_taxon_name: s, object_taxon_name: o, type: 'TaxonNameRelationship::Iczn::Invalidating::Usage::FamilyGroupNameForm')
-                elsif (o.rank_string =~ /Species/  && shas == o.cached_secondary_homonym_alternative_spelling) ||
-                    (o.rank_string =~ /Genus/  && s.cached_primary_homonym_alternative_spelling == o.cached_primary_homonym_alternative_spelling)
+                elsif (o.rank_string =~ /Species/  && shas == o.cached_secondary_homonym_alternative_spelling && r2 == 1) ||
+                    (o.rank_string =~ /Genus/  && s.cached_primary_homonym_alternative_spelling == o.cached_primary_homonym_alternative_spelling && r2 == 1)
                   combinations += 1
+                  byebug if s.type != 'Protonym'
                   genus = s.original_genus
                   subgenus = s.original_subgenus
                   species = s.original_species
                   subspecies = s.original_subspecies
-                  vname = s.cached_original_combination.gsub('<i>', '').gsub('</i>', '')
+                  vname = s.cached_original_combination.to_s.gsub('<i>', '').gsub('</i>', '')
                   s.original_genus_relationship.destroy unless genus.blank?
                   s.original_subgenus_relationship.destroy unless subgenus.blank?
                   s.original_species_relationship.destroy unless species.blank?
@@ -2177,8 +2188,6 @@ namespace :tw do
                   s.rank_class = nil
                   s.type = 'Combination'
                   s = s.becomes(Combination)
-                  #s.destroy
-                  #s = Combination.new
                   s.genus = genus unless genus.nil?
                   s.subgenus = subgenus unless subgenus.nil?
                   s.species = species unless species.nil?
@@ -2194,13 +2203,23 @@ namespace :tw do
                     s.genus = o
                   end
                   s.save
-                  unless s.valid?
+                  if !s.valid?
                     s = Protonym.find(s.id)
                     TaxonNameRelationship.create!(subject_taxon_name: s, object_taxon_name: o, type: 'TaxonNameRelationship::Iczn::Invalidating')
          #           c1 = Combination.match_exists?(s.get_full_name, genus: s.genus.try(:id), subgenus: s.subgenus.try(:id), species: s.species.try(:id), subspecies: s.subspecies.try(:id))
          #           c1 = Combination.matching_protonyms(s.get_full_name_html, genus: s.genus.try(:id), subgenus: s.subgenus.try(:id), species: s.species.try(:id), subspecies: s.subspecies.try(:id)).first if c1.blank?
          #           byebug if c1.blank?
          #           s = c1
+                  else
+                    TaxonNameRelationship.where(project_id: $project_id, subject_taxon_name_id: s.id).with_type_contains('Combination').each do |z|
+                      z.object_taxon_name.verbatim_name = z.object_taxon_name.cached if z.object_taxon_name.type = 'Combination' && z.object_taxon_name.verbatim_name.blank?
+                      z.subject_taxon_name_id = o.id
+                      z.save
+                      z.subject_taxon_name.save
+#                      byebug if !z.valid?
+#                      byebug if !z.object_taxon_name.valid?
+                      fixed += 1
+                    end
                   end
 
 
@@ -2217,31 +2236,38 @@ namespace :tw do
         end
 #end
 
-#        print "\nHandling Invalid relationships: fix combination relationships\n"
+
+        print "\nHandling Invalid relationships: cached valid\n"
+        i = 0
+        fixed = 0
+        tn = TaxonName.where(project_id: $project_id)
+        tn.each do |t|
+          i += 1
+          print "\r#{i}    wrong cached valid fixed: #{fixed}   "
+          if t.cached_valid_taxon_name_id != t.id && t.valid_taxon_name.type == 'Combination'
+            t.save
+            fixed += 1 if t.valid_taxon_name.type == 'Protonym'
+            byebug if t.valid_taxon_name.type == 'Combination'
+          end
+        end
+
+
+#        print "\nHandling Invalid relationships: wrong combination relationships\n"
 #        i = 0
 #        fixed = 0
 #        tr = TaxonNameRelationship.where(project_id: $project_id).with_type_contains('Combination')
 #        tr.each do |t|
 #          i += 1
-#          print "\r#{i}    Fixes applied: #{fixed}   "
+#          print "\r#{i}    Wrong combinations found: #{fixed}   "
 #          s = t.subject_taxon_name
 #          if s.type == 'Combination'
-#            byebug
-#            if !s.subspecies.nil?
-#              t.subject_taxon_name = s.subspecies
-#            elsif !s.species.nil?
-#              t.subject_taxon_name = s.species
-#            elsif !s.subgenus.nil?
-#              t.subject_taxon_name = s.subgenus
-#            elsif !s.genus.nil?
-#              t.subject_taxon_name = s.genus
-#            end
+#            print "\nWrong combination id: #{t.object_taxon_name_id}\n"
 
-#            t.save
-#            fixed += 1 if t.valid?
+#            fixed += 1
 #          end
 #        end
-        end
+
+      end
 
     end
   end
